[refactor] Others Page -> User Home으로 네이밍 변경
