+++ app/src/main/res/layout/item_user_home_challenge.xml
